feat: 라우터 생성, 컨트롤러 연결, 코드 리팩터링 및 오누락사항 수정
