+++ server/src/controllers/user.ts
@@ -1,14 +1,13 @@
 import type { Context } from 'koa';
-import twilio from 'twilio';
 import bcrypt from 'bcrypt';
 import { validate } from 'class-validator';
 import type { ValidationError } from 'class-validator';
 
 import User from 'src/models/user';
-import { CreateUserDto, SendSMSVerificationCodeDto } from 'typings/auth.dto';
+import { CreateUserDto } from 'typings/auth.dto';
 import { CreateUserProps } from 'typings/auth';
 
-export const verifyExistenceUsername = async (ctx: Context) => {
+export const verifyExistenceUsername = async (ctx: Context): Promise<void> => {
   const { username } = ctx.params;
 
   const exUser = await User.findOne({
@@ -24,61 +23,7 @@ export const verifyExistenceUsername = async (ctx: Context) => {
   ctx.body = { message: '사용 가능한 username 입니다.' };
 };
 
-export const sendSMSVerificationCode = async (ctx: Context) => {
-  const { phoneNumber } = ctx.request.body;
-
-  const sendSMSVerificationCodeDto = new SendSMSVerificationCodeDto();
-  sendSMSVerificationCodeDto.phoneNumber = phoneNumber;
-
-  const validationErrors: ValidationError[] = await validate(sendSMSVerificationCodeDto);
-  if (validationErrors.length > 0) {
-    ctx.status = 400;
-    ctx.body = { message: validationErrors };
-    return;
-  }
-
-  const max = 999_999;
-  const randomNum = Math.ceil(Math.random() * max);
-  const verificationCodeStr = randomNum.toString().padStart(6, '0'); // '000001' ~ '999999'
-
-  if (!ctx.session) throw new Error('ctx 에 session 프로퍼티가 존재하지 않습니다.');
-  ctx.session.verificationCode[phoneNumber] = verificationCodeStr;
-
-  const accountSid = process.env.TWILIO_ACCOUNT_SID;
-  const authToken = process.env.TWILIO_AUTH_TOKEN;
-  const client = twilio(accountSid, authToken);
-
-  const recipient = client.messages
-    .create({
-      body: '',
-      from: process.env.TWILIO_FROM,
-      to: phoneNumber,
-    })
-    .then((info) => info.to)
-    .catch((err) => console.error(err));
-
-  ctx.status = 200;
-  ctx.body = { message: `${recipient} 번호로 인증번호가 전송되었습니다.` };
-};
-
-export const checkSMSVerificationCode = async (ctx: Context) => {
-  const { phoneNumber, code } = ctx.request.body;
-
-  if (!ctx.session) throw new Error('ctx 에 session 프로퍼티가 존재하지 않습니다.');
-  const sentCode = ctx.session.verificationCode[phoneNumber];
-  if (!sentCode)
-    throw new Error(`${phoneNumber} 번호로 저장했던 인증 코드가 만료되었거나 존재하지 않습니다.`);
-
-  if (code !== sentCode) {
-    ctx.status = 400;
-    ctx.body = { message: '잘못된 인증번호입니다.' };
-    return;
-  }
-  ctx.status = 202;
-  ctx.body = { message: '인증번호 확인 완료.' };
-};
-
-export const createUser = async (ctx: Context) => {
+export const createUser = async (ctx: Context): Promise<void> => {
   const createUserFields: CreateUserProps = ctx.request.body;
 
   const createUserDto = new CreateUserDto();
