+++ server/src/routes/hot.ts
@@ -1,0 +1,8 @@
+import Router from 'koa-router';
+import hotReloadingSSE from 'src/controllers/hot';
+
+const router = new Router();
+
+router.get('/', hotReloadingSSE);
+
+export default router;
