+++ server/src/routes/sms.ts
@@ -1,0 +1,9 @@
+import Router from 'koa-router';
+import { sendSMSVerificationCode, checkSMSVerificationCode } from 'src/controllers/sms';
+
+const router = new Router();
+
+router.post('/', sendSMSVerificationCode);
+router.post('/match', checkSMSVerificationCode);
+
+export default router;
