+++ server/src/routes/user.ts
@@ -1,0 +1,9 @@
+import Router from 'koa-router';
+import { verifyExistenceUsername, createUser } from 'src/controllers/user';
+
+const router = new Router();
+
+router.get('/:username/exist', verifyExistenceUsername);
+router.post('/', createUser);
+
+export default router;
